+++ lib/PageHelpers/animated_page_dragger.dart
@@ -3,38 +3,42 @@ import 'dart:ui';
 import 'package:flutter/material.dart';
 import 'package:liquid_swipe/Helpers/Helpers.dart';
 import 'package:liquid_swipe/Helpers/slide_update.dart';
-import 'package:liquid_swipe/Provider/iamariderprovider.dart';
+import 'package:liquid_swipe/Provider/LiquidProvider.dart';
 
+/// Internal Class
+///
 /// This class provides the animation controller
-/// used when then user stops dragging and page
+/// used when user stops dragging and page
 /// reveal is not completed.
-
 class AnimatedPageDragger {
+  ///SlideDirection LTR, RTL or none
   final SlideDirection slideDirection;
 
-  //This variable tells that whether we have to open or close the page reveal.
+  ///Current transition goal, either close the page or reveal it
   final TransitionGoal transitionGoal;
 
-  //Animation controller
-  AnimationController completionAnimationController;
+  ///Animation controller for Completing the Animation when user is Done with dragging
+  late AnimationController completionAnimationController;
 
-  //Constructor
+  ///Constructor
   AnimatedPageDragger({
-    this.slideDirection,
-    this.transitionGoal,
-    double slidePercentVer,
-    double slidePercentHor,
-    @required IAmARiderProvider slideUpdateStream,
-    TickerProvider vsync,
+    required this.slideDirection,
+    required this.transitionGoal,
+    required double slidePercentVer,
+    required double slidePercentHor,
+    required LiquidProvider slideUpdateStream,
+    required TickerProvider vsync,
   }) {
     final startSlidePercentHor = slidePercentHor;
     final startSlidePercentVer = slidePercentVer;
-    double endSlidePercentHor, endSlidePercentVer;
+    double? endSlidePercentHor, endSlidePercentVer;
     Duration duration;
 
     //We have to complete the page reveal
     if (transitionGoal == TransitionGoal.open) {
-      endSlidePercentHor = endSlidePercentVer = 1.0;
+      endSlidePercentHor = 1.0;
+
+      endSlidePercentVer = slideUpdateStream.positionSlideIcon;
 
       final slideRemaining = 1.0 - slidePercentHor;
       //Standard value take for drag velocity to avoid complex calculations.
@@ -71,8 +75,8 @@ class AnimatedPageDragger {
         //Adding to slide update stream
         slideUpdateStream.updateSlide(SlideUpdate(
           slideDirection,
-          slidePercent,
-          slidePercentVer,
+          slidePercent!,
+          slidePercentVer!,
           UpdateType.animating,
         ));
       })
@@ -90,12 +94,12 @@ class AnimatedPageDragger {
       });
   }
 
-  //This method is used to run animation Controller
+  ///This method is used to run animation Controller in forward
   void run() {
     completionAnimationController.forward(from: 0.0);
   }
 
-  //This method is used to dispose animation controller
+  ///This method is used to dispose animation controller
   void dispose() {
     completionAnimationController.dispose();
   }